move provisioner guest helper functions into common dir.
