+++ common/guestexec/elevated_test.go
@@ -1,4 +1,4 @@
-package provisioner
+package guestexec
 
 import (
 	"regexp"
