+++ common/guestexec/guest_commands.go
@@ -1,4 +1,4 @@
-package provisioner
+package guestexec
 
 import (
 	"fmt"
