+++ common/guestexec/guest_commands_test.go
@@ -1,4 +1,4 @@
-package provisioner
+package guestexec
 
 import (
 	"testing"
